refactor(infra): bedrock配下を機能フォルダに再配置し、重複/大小文字衝突を解消
- ContentProcessing/index.ts と QueryOptimization/index.ts に実装を集約
- 旧 contentProcessing.ts / queryOptimization.ts を削除
- バレルと参照（UseCase/テスト）を新構成へ更新

Refs: #9
