+++ src/shared/infrastructure/external/bedrock/QueryOptimization/index.ts
