+++ src/shared/infrastructure/external/bedrock/index.ts
@@ -1,1 +1,2 @@
-export * from "./contentProcessing";
+export * from "./ContentProcessing";
+export * from "./QueryOptimization";
